unit test

+++ zeeArduino.cpp
@@ -1,4 +1,5 @@
 #include "zeeArduino.h"
+#include "Arduino.h"
 
 zeeArduino::zeeArduino(unsigned int updateInterval, unsigned int intervalRunTime)
   :_updateInterval(updateInterval), _intervalRunTime(intervalRunTime)
@@ -19,22 +20,22 @@ void zeeArduino::delayMicroseconds(unsigned int us)
   Arduino_h::delayMicroseconds(us);
 }
 
-unsigned long zeeArduino::pulseIn(uint8_t pin, uint8_t state, unsigned long timeout)
+unsigned long zeeArduino::pulseIn(unsigned int pin, unsigned int state, unsigned long timeout)
 {
    return Arduino_h::pulseIn(pin, state, timeout);
 }
 
-unsigned long zeeArduino::pulseInLong(uint8_t pin, uint8_t state, unsigned long timeout)
+unsigned long zeeArduino::pulseInLong(unsigned int pin, unsigned int state, unsigned long timeout)
 {
   return Arduino_h::pulseInLong(pin, state, timeout);
 }
 
-void zeeArduino::digitalWrite(uint8_t pin, uint8_t highlow)
+void zeeArduino::digitalWrite(unsigned int pin, unsigned int highlow)
 {
   Arduino_h::digitalWrite(pin, highlow);
 }
 
-int zeeArduino::digitalRead(uint8_t pin)
+int zeeArduino::digitalRead(unsigned int pin)
 {
   return Arduino_h::digitalRead(pin);
 }
@@ -44,7 +45,7 @@ unsigned long zeeArduino::millis()
   return Arduino_h::millis();
 }
 
-void zeeArduino::pinMode(byte pin, byte direction)
+void zeeArduino::pinMode(unsigned int pin, unsigned int direction)
 {
-  Arduino_h::pinMode(pin, direction);
+  Arduino_h::pinMode((uint8_t)pin, (uint8_t)direction);
 }
